machine-readable-data-licenses/1.0.md
The wording suggests that once a license has been published into the Registry, it is "immutable". This makes sense but it isn't clear how a license is removed from the Registry, should that ever be needed.

Needs clarity on wording to explain what happens when a license is removed, revoked or amended.