Suggested changes to member-identity-digital-certificates/1.0.md
Need to add in how long server certificates last because it's not clear in the documentation.

Have removed where it says that certificates do not have to be checked for revocation, because that's implying a lack of risk when it does exist.

Have also removed the specification of ACME for automatic renewal because even in the case of ACME, it's a manual process to install a certificate and we haven't yet agreed that ACME is the correct process.